Make the upstart user declarations work
Fixes #9

I also made the default for user and group undef. I did this because
Upsart (and many other init systems) will choose these values for you if
left unspecified, and they'll typically know better than this module.
For instance: "If this stanza is unspecified, the primary group of the
user specified in the setuid block is used" (Upstart) and "If no group
is set, the default group of the user is chosen" (systemd).