Rely soley on $PIDFILE in sysv_debian do_stop()
Using a --name filter when stopping a service is unreliable for two
reasons. First, the name is limited to 15 characters which is very
limiting since we are passing $DAEMON which will be a full path to an
executable. Second, interpreted scripts must match their interpreter and
not the full path to the script.

Instead, rely soley on $PIDFILE which we create in do_start() and should
always exist when the service is running.